requirements.txt
asgiref==3.5.2
attrs==22.1.0
distlib==0.3.6
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
inflection==0.5.1
iniconfig==1.1.1
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
virtualenv==20.16.5
django-cors-headers
